Mon, Sep 30, 2024, 8:58 AM +02:00

+++ bitcoin.py
@@ -10,6 +10,6 @@ o = response.json()
 
 o = o["bpi"]["USD"]["rate"]
 
-o = float(o)
+o = int(o)
 
 print(o * x)
